Add a new parsing regression test.
This is intended as a fast check that all tags, including MakerNotes, can be
parsed.  It is not intended to replace the much more complete tests in
libexif-testsuite, but rather add some basic parsing test coverage without
having to download and configure a separate repository. As more tags are
identified and supported in the library in the future, the golden test files
will need to be updated to match.

The input files cover all four major MakerNote types supported by libexif and a
large proportion of the variants within those types.  The test files have had
their JPEG image data stripped out to make them smaller (which also makes them
noncompliant) but it's good enough for libexif to read the tags as well as some
other EXIF tools that don't care about the image.

Patch from Google.