Added test-extract.c and a new regression test extract-parse.sh.
This extracts the EXIF tags from an image then compares the parsed value
of the extracted tags with those of the original file. This ensures that
the tags are written properly, without change in tag data. The MakerNote
tag sometimes has a harmless, slight difference in size because of
padding being removed.

However, in developing this test, I found that the Olympus variant 4
MakerNote has a huge size difference. This might be harmless (there
might just be a lot of padding removed) but it's also possible that
these MakerNotes aren't being properly parsed.  This discrepancy should
be investigated.

The exif_data_save_data() function is also returning some JPEG markers
at the end of the buffer which I wasn't expecting.  This also should be
investigated.

The test is enabled anyway in the meantime to reduce the chance of
regressions in the remaining tags.