test: Do not fail tests when there are whitespace differences.
For some reason I cannot reproduce the whitespace on one
of the values, where the whitespace ammount changes from
the expected for no reason. I think this is not relevant,
so we make the diff ignore whitespace amount. Notice it
ignores white-space change not ignore all white space
changes. This is more conservative to make sure I am not
breaking anything.