Use Windows localtime_s() if available
If available, use localtime_s() instead of localtime().

As localtime_s() is part of the Windows API, the compile
check with localtime_s() should only work for those systems.

Note we are not using AC_CANONICAL_HOST and then checking the
$host_os here. Either localtime_s() is available, then we
can use it. Or localtime_s() is not available, then we cannot
use it.

All of that is independent of what exactly the $host_os value
actually is, and whether we have listed all appropriate values
in the list of matches.

Closes: https://github.com/libexif/libexif/issues/57
Closes: https://github.com/libexif/libexif/pull/66