check for diff and diff -u before using it
Before actually running diff, check that diff is actually
available. And when it is possible, even use "diff -u".

And if diff is not available at all, fall back to cmp.
cmp does not produce nice verbose output, but it has
the same exit values as diff.

Oh, and we try to avoid "if ! condition" in the shell script,
as there allegedly are some corner cases where that is
problematic.