CPP macros beginning/ending with underscores are reserved
C preprocessor macros beginning/ending with underscores are
reserved for system use, so application software macros are
supposed to be named differently.

For a header file foo-bar.h, this consistently use a macro
name LIBEXIF_FOO_BAR_H for the usual

    #ifndef LIBEXIF_FOO_BAR_H
    #define LIBEXIF_FOO_BAR_H
    ... content of the header file ...
    #endif /* LIBEXIF_FOO_BAR_H */

Found using "clang -Weverything".

This should work, unless when run with an ancient C preprocessor
which might only the first 8 or 10 characters to distinguish
macro names and ignores the remainder of the name.