cast 0 to enum type where enum type is expected
There are a few list terminators which use a 0 value for an
enum field of a struct. This casts these 0 values to the
respective enum type to avoid some warnings with "clang
-Weverything".

Every C compiler supporting enums should be fine with
getting an value of the enum type to initialize a struct
field of that enum type.