There is no need for "default:" if all cases are covered
There is no need for "default:" if all enum cases are covered, and
some compilers can even generate a warning about that.

However, there is still a need to handle the case where someone has
passed an integer value as the enum for which the enum definition and
thus the switch has no special value.

We do that outside the switch, and abort() with error message.