Reinstate the code coverage build to CI
The results is pushed to Coveralls at
https://coveralls.io/github/libexif/libexif  The coverage figures there
haven't been updated since the Travis builds stopped over 3 years ago
and the coverage has gone down a few percent in that time.

The host OS is the older 22.04 Ubuntu release because that's the last
LTS release that supports the glibc malloc hooks that failmalloc uses to
inject memory errors. Another approach will be needed once that version
is no longer available.