Add tests using test-mnote
This was previously built but never run. The check-mnote.sh script now
runs it against all test data files in order to look for gross errors,
such as those causing SIGSEGV or SIGFPE, or more when compiled using
features like ASAN and UBSAN.

Also, fix test-mnote to add a missing free and to not pass NULL values
to printf.