Remove failmalloc from the coverage build
When coverage support is enabled, it causes thousands of mallocs to
occur before the test programs even get a chance to start running. This
destroys assumptions made in check-failmalloc.sh about allocations and
therefore causes no additional code coverage. Also, configure the
sanitize build to abort on a UBSAN error.